pyproject.toml
[project]
name = "volatilityengine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.3.0",
    "fastapi>=0.133.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "yfinance>=1.2.0",
]
